chore: move addressForm to containers folder
it is a component that handles state, so it should be a container
the theme prop was also added so it can have its own styling

the states.js file, as it is used in more than one place, was moved to
the helpers folder
