+++ webapp/src/pages/Shipping/style.css
@@ -78,18 +78,3 @@
 .visible {
   display: block;
 }
-
-.hidden {
-  display: none;
-}
-
-.addressForm {
-  background-color: var(--addressForm-background-color);
-  border: 2px dashed var(--addressForm-border-color);
-  padding: 33px 24px;
-  margin-top: 24px;
-}
-
-.cancelButton {
-  margin-right: 26px;
-}
