+++ webapp/src/theme-pagarme/address-form/properties.css
@@ -1,0 +1,11 @@
+@import "../colors.css";
+@import "../typography.css";
+
+:root {
+  --addressForm-background-color: var(--color-grey-80);
+  --addressForm-border-color: var(--color-grey-60);
+  --page-title-font-family: var(--primary-font-family);
+  --page-title-font-size: var(--font-size-small);
+  --page-title-color: var(--color-white);
+  --page-title-margin: 6px 0;
+}
